PET-92 refactor : TestSpringBootApplicationConfig 위치 testFixture로 변경
