+++ Common-Module/src/testFixtures/java/com/pawith/TestSpringBootApplicationConfig.java
@@ -1,7 +1,9 @@
-package com.pawith.authmodule.adaptor.api;
+package com.pawith;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.test.context.TestConfiguration;
 
+@TestConfiguration
 @SpringBootApplication
 public class TestSpringBootApplicationConfig {
 }
