--- Domain-Module/User-Module/User-Presentation/src/test/java/com/pawith/userpresentation/TestSpringBootApplicationConfig.java
@@ -1,7 +1,0 @@
-package com.pawith.userpresentation;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TestSpringBootApplicationConfig {
-}
